move ZlibService and Exceptions to more appropriate namepsace
