--- SIT.Manager.Avalonia/Classes/Exceptions/AccountNotFoundException.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SIT.Manager.Avalonia.Classes.Exceptions
-{
-    public class AccountNotFoundException : TarkovException
-    {
-    }
-}
