+++ SIT.Manager.Avalonia/Exceptions/AccountNotFoundException.cs
@@ -1,0 +1,6 @@
+﻿namespace SIT.Manager.Avalonia.Exceptions
+{
+    public class AccountNotFoundException : TarkovException
+    {
+    }
+}
