+++ SIT.Manager.Avalonia/Exceptions/IncorrectServerPasswordException.cs
@@ -1,0 +1,5 @@
+﻿namespace SIT.Manager.Avalonia.Exceptions;
+
+public class IncorrectServerPasswordException : TarkovException
+{
+}
