+++ SIT.Manager.Avalonia/Exceptions/TarkovException.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace SIT.Manager.Avalonia.Exceptions;
+
+public class TarkovException : Exception
+{
+}
